Move btpoperator controller tests to separate files
