+++ controllers/btpoperator_controller_configuration_test.go
@@ -1,0 +1,19 @@
+package controllers
+
+import (
+	"time"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+var _ = Describe("BTP Operator controller - configuration", func() {
+	Context("When the ConfigMap is present", func() {
+		It("should adjust configuration settings in the operator accordingly", func() {
+			GinkgoWriter.Println("--- PROCESS:", GinkgoParallelProcess(), "---")
+			cm := initConfig(map[string]string{"ProcessingStateRequeueInterval": "10s"})
+			reconciler.reconcileConfig(cm)
+			Expect(ProcessingStateRequeueInterval).To(Equal(time.Second * 10))
+		})
+	})
+})
