+++ controllers/btpoperator_controller_deprovisioning_test.go
@@ -1,0 +1,126 @@
+package controllers
+
+import (
+	"github.com/kyma-project/btp-manager/api/v1alpha1"
+	"github.com/kyma-project/module-manager/pkg/types"
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+const (
+	instanceName = "my-service-instance"
+	bindingName  = "my-service-binding"
+)
+
+var _ = Describe("BTP Operator controller - deprovisioning", func() {
+	var cr *v1alpha1.BtpOperator
+
+	Describe("Deprovisioning without force-delete label", func() {
+
+		BeforeEach(func() {
+			GinkgoWriter.Println("--- PROCESS:", GinkgoParallelProcess(), "---")
+			secret, err := createCorrectSecretFromYaml()
+			Expect(err).To(BeNil())
+			Expect(k8sClient.Patch(ctx, secret, client.Apply, client.ForceOwnership, client.FieldOwner(operatorName))).To(Succeed())
+			cr = createBtpOperator()
+			Expect(k8sClient.Create(ctx, cr)).To(Succeed())
+			Eventually(updateCh).Should(Receive(matchState(types.StateReady)))
+			btpServiceOperatorDeployment := &appsv1.Deployment{}
+			Expect(k8sClient.Get(ctx, client.ObjectKey{Name: DeploymentName, Namespace: kymaNamespace}, btpServiceOperatorDeployment)).Should(Succeed())
+		})
+
+		AfterEach(func() {
+			Expect(k8sClient.Get(ctx, client.ObjectKey{Namespace: defaultNamespace, Name: btpOperatorName}, cr)).To(Succeed())
+			cr.SetLabels(map[string]string{forceDeleteLabelKey: "true"})
+			Expect(k8sClient.Update(ctx, cr)).To(Succeed())
+			Eventually(func() (bool, error) {
+				err := k8sClient.Get(ctx, client.ObjectKey{Namespace: defaultNamespace, Name: btpOperatorName}, cr)
+				return cr.Labels[forceDeleteLabelKey] == "true", err
+			}).WithTimeout(k8sOpsTimeout).WithPolling(k8sOpsPollingInterval).Should(BeTrue())
+			Eventually(updateCh).Should(Receive(matchDeleted()))
+			deleteSecret := &corev1.Secret{}
+			Expect(k8sClient.Get(ctx, client.ObjectKey{Namespace: kymaNamespace, Name: SecretName}, deleteSecret)).To(Succeed())
+			Expect(k8sClient.Delete(ctx, deleteSecret)).To(Succeed())
+		})
+
+		It("Delete should fail because of existing instances and bindings", func() {
+			_ = createResource(instanceGvk, kymaNamespace, instanceName)
+			ensureResourceExists(instanceGvk)
+
+			_ = createResource(bindingGvk, kymaNamespace, bindingName)
+			ensureResourceExists(bindingGvk)
+
+			Expect(k8sClient.Get(ctx, client.ObjectKey{Namespace: defaultNamespace, Name: btpOperatorName}, cr)).To(Succeed())
+			Expect(k8sClient.Delete(ctx, cr)).To(Succeed())
+
+			Eventually(updateCh).Should(Receive(matchReadyCondition(types.StateDeleting, metav1.ConditionFalse, ServiceInstancesAndBindingsNotCleaned)))
+		})
+
+	})
+
+	Describe("Deprovisioning with force-delete label", func() {
+		var siUnstructured, sbUnstructured *unstructured.Unstructured
+
+		BeforeEach(func() {
+			GinkgoWriter.Println("--- PROCESS:", GinkgoParallelProcess(), "---")
+			secret, err := createCorrectSecretFromYaml()
+			Expect(err).To(BeNil())
+			Expect(k8sClient.Patch(ctx, secret, client.Apply, client.ForceOwnership, client.FieldOwner(operatorName))).To(Succeed())
+			cr = createBtpOperator()
+			cr.SetLabels(map[string]string{forceDeleteLabelKey: "true"})
+			Expect(k8sClient.Create(ctx, cr)).To(Succeed())
+			Eventually(updateCh).Should(Receive(matchState(types.StateReady)))
+			btpServiceOperatorDeployment := &appsv1.Deployment{}
+			Expect(k8sClient.Get(ctx, client.ObjectKey{Name: DeploymentName, Namespace: kymaNamespace}, btpServiceOperatorDeployment)).Should(Succeed())
+
+			siUnstructured = createResource(instanceGvk, kymaNamespace, instanceName)
+			ensureResourceExists(instanceGvk)
+
+			sbUnstructured = createResource(bindingGvk, kymaNamespace, bindingName)
+			ensureResourceExists(bindingGvk)
+		})
+
+		AfterEach(func() {
+			deleteSecret := &corev1.Secret{}
+			Expect(k8sClient.Get(ctx, client.ObjectKey{Namespace: kymaNamespace, Name: SecretName}, deleteSecret)).To(Succeed())
+			Expect(k8sClient.Delete(ctx, deleteSecret)).To(Succeed())
+		})
+
+		It("soft delete (after timeout) should succeed", func() {
+			reconciler.Client = newTimeoutK8sClient(reconciler.Client)
+			setFinalizers(siUnstructured)
+			setFinalizers(sbUnstructured)
+			Expect(k8sClient.Get(ctx, client.ObjectKey{Namespace: defaultNamespace, Name: btpOperatorName}, cr)).To(Succeed())
+			Expect(k8sClient.Delete(ctx, cr)).To(Succeed())
+			Eventually(updateCh).Should(Receive(matchReadyCondition(types.StateDeleting, metav1.ConditionFalse, HardDeleting)))
+			Eventually(updateCh).Should(Receive(matchReadyCondition(types.StateDeleting, metav1.ConditionFalse, SoftDeleting)))
+			Eventually(updateCh).Should(Receive(matchDeleted()))
+			doChecks()
+		})
+
+		It("soft delete (after hard deletion fail) should succeed", func() {
+			reconciler.Client = newErrorK8sClient(reconciler.Client)
+			setFinalizers(siUnstructured)
+			setFinalizers(sbUnstructured)
+			Expect(k8sClient.Get(ctx, client.ObjectKey{Namespace: defaultNamespace, Name: btpOperatorName}, cr)).To(Succeed())
+			Expect(k8sClient.Delete(ctx, cr)).To(Succeed())
+			Eventually(updateCh).Should(Receive(matchReadyCondition(types.StateDeleting, metav1.ConditionFalse, SoftDeleting)))
+			Eventually(updateCh).Should(Receive(matchDeleted()))
+			doChecks()
+		})
+
+		It("hard delete should succeed", func() {
+			reconciler.Client = k8sClientFromManager
+			Expect(k8sClient.Get(ctx, client.ObjectKey{Namespace: defaultNamespace, Name: btpOperatorName}, cr)).To(Succeed())
+			Expect(k8sClient.Delete(ctx, cr)).Should(Succeed())
+			Eventually(updateCh).Should(Receive(matchReadyCondition(types.StateDeleting, metav1.ConditionFalse, HardDeleting)))
+			Eventually(updateCh).Should(Receive(matchDeleted()))
+			doChecks()
+		})
+	})
+})
